+++ src/main/java/edu/uc/gamelibrarymanager/WebSecurityConfig.java
@@ -1,6 +1,8 @@
 package edu.uc.gamelibrarymanager;
 
 import com.google.api.client.http.HttpMethods;
+import edu.uc.gamelibrarymanager.security.AuthenticationProvider;
+import edu.uc.gamelibrarymanager.security.FirebaseAuthenticationFilter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
